requirements.txt
Django (2.0)
djangorestframework (3.7.3)
olefile (0.44)
Pillow (4.3.0)
pip (9.0.1)
pytz (2017.3)
setuptools (28.8.0)
